Abstract event subscriber for projections
